fn: new agent resource tracker metrics
New metrics for agent resource tracker: CpuUsed, CpuAvail,
MemUsed, MemAvail.

+++ api/agent/call.go
@@ -427,7 +427,7 @@ func (a *agent) GetCall(opts ...CallOpt) (Call, error) {
 	}
 
 	mem := c.Memory + uint64(c.TmpFsSize)
-	if !a.resources.IsResourcePossible(mem, uint64(c.CPUs), c.Type == models.TypeAsync) {
+	if !a.resources.IsResourcePossible(mem, c.CPUs, c.Type == models.TypeAsync) {
 		return nil, models.ErrCallResourceTooBig
 	}
